apps/worker/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
redis==5.2.1
pydantic-settings==2.8.1
google-genai==1.14.0
google-cloud-storage==2.19.0
google-cloud-pubsub==2.25.0
python-docx==1.1.2
pytest==8.3.5
